pyproject.toml
[tool.poetry]
name = "nuvemfiscal"
version = "0.1.0"
description = ""
authors = ["Leonardo Gregianin <leogregianin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
httpx = "^0.23.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
